requirement.txt
discord.py
discord.py[voice]
youtube_dl
pafy
